Add Interface for `BackendManager` to allow easier testing
